backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appraisal360-backend"
version = "0.1.0"
description = "FastAPI backend for Appraisal360"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic-settings==2.5.2",
  "sqlalchemy==2.0.35",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "boto3==1.35.24",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "email-validator==2.2.0",
]

[project.scripts]
appraisal360-backend = "src.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
